+++ src/aioway/io/frames/__init__.py
@@ -2,4 +2,3 @@
 
 from .chunks import *
 from .frames import *
-from .views import *
